Updated README; fixes; set version to 1.0.0
Updated README to latest changes and completed version history in
README.md.
Set version to 1.0.0 in Project.toml.
When reading in MCM database files in function get_photlabel, don't use
column names to address columns in the dataframes, but index 1 for
reaction numbers, and end for TUV reaction labels. This allows an
additional column with scaling factors for MCMv3.2 and older versions
needed elsewhere.